Update update.php
Fixes #93

+++ update.php
@@ -116,6 +116,11 @@ if (Input::get('action') === 'editprofile') {
         Redirect::to('pireps.php');
     }
 } elseif (Input::get('action') === 'edituser') {
+    if (!$user->hasPermission('usermanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     $perms = Json::decode($user->getUser(Input::get('id'))["permissions"]);
     $perms["admin"] = Input::get("admin");
 
@@ -131,6 +136,10 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'User Edited Successfully!');
     Redirect::to('admin.php?page=usermanage');
 } elseif (Input::get('action') === 'deluser') {
+    if (!$user->hasPermission('usermanage')) {
+        Redirect::to('home.php');
+        die();
+    }
 
     try {
         $user->update(array(
@@ -143,6 +152,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'User deleted successfully!');
     Redirect::to('admin.php?page=usermanage');
 } elseif (Input::get('action') === 'editstaffmember') {
+    if (!$user->hasPermission('staffmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     $permissions = Permissions::getAll();
     $finalpermissions = array('admin' => 1);
     foreach ($permissions as $permission => $data) {
@@ -168,6 +182,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'Staff Member Edited Successfully!');
     Redirect::to('admin.php?page=staffmanage');
 } elseif (Input::get('action') === 'declineapplication') {
+    if (!$user->hasPermission('recruitment')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     try {
         $user->update(array(
             'status' => 3,
@@ -180,6 +199,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'Application Declined Successfully');
     Redirect::to('admin.php?page=recruitment');
 } elseif (Input::get('action') === 'acceptapplication') {
+    if (!$user->hasPermission('recruitment')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     try {
         $user->update(array(
             'status' => 1
@@ -191,18 +215,38 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'Application Accepted Successfully!');
     Redirect::to('admin.php?page=recruitment');
 } elseif (Input::get('action') === 'acceptpirep') {
+    if (!$user->hasPermission('pirepmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Pirep::accept(Input::get('accept'));
     Session::flash('success', 'PIREP Accepted Successfully!');
     Redirect::to('admin.php?page=pirepmanage');
 } elseif (Input::get('action') === 'declinepirep') {
+    if (!$user->hasPermission('pirepmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Pirep::decline(Input::get('decline'));
     Session::flash('success', 'PIREP Declined Successfully');
     Redirect::to('admin.php?page=pirepmanage');
 } elseif (Input::get('action') === 'deletearticle') {
+    if (!$user->hasPermission('newsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     News::archive(Input::get('delete'));
     Session::flash('success', 'News Item Archived Successfully! ');
     Redirect::to('admin.php?page=newsmanage');
 } elseif (Input::get('action') === 'editarticle') {
+    if (!$user->hasPermission('newsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     News::edit(Input::get('id'), array(
         'subject' => Input::get('title'),
         'content' => Input::get('content')
@@ -210,6 +254,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'News Article Edited Successfully! ');
     Redirect::to('admin.php?page=newsmanage');
 } elseif (Input::get('action') === 'newarticle') {
+    if (!$user->hasPermission('newsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     News::add(array(
         'subject' => Input::get('title'),
         'content' => Input::get('content'),
@@ -218,10 +267,20 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'News Article Added Successfully! ');
     Redirect::to('admin.php?page=newsmanage');
 } elseif (Input::get('action') === 'deleteaircraft') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Aircraft::archive(Input::get('delete'));
     Session::flash('success', 'Aircraft Archived Successfully! ');
     Redirect::to('admin.php?page=opsmanage&section=fleet');
 } elseif (Input::get('action') === 'addaircraft') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Aircraft::add(Input::get('aircraftselect'), Rank::nameToId(Input::get('rank')), Input::get('livery'));
     Session::flash('success', 'Aircraft Added Successfully! ');
     Redirect::to('admin.php?page=opsmanage&section=fleet');
@@ -233,18 +292,38 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('successrecent', 'PIREPs Setup Successfully! You can now File PIREPs.');
     Redirect::to('pireps.php?page=new');
 } elseif (Input::get('action') === 'addroute') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Route::add(array(Input::get('fltnum'), Input::get('dep'), Input::get('arr'), Time::strToSecs(Input::get('duration')), Aircraft::nameToId(Input::get('aircraft'))));
     Session::flash('success', 'Route Added Successfully!');
     Redirect::to('admin.php?page=opsmanage&section=routes');
 } elseif (Input::get('action') === 'deleteroute') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Route::delete(Input::get('deleteroute'));
     Session::flash('success', 'Route Removed Successfully!');
     Redirect::to('admin.php?page=opsmanage&section=routes');
 } elseif (Input::get('action') === 'addrank') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     Rank::add(Input::get('name'), Time::hrsToSecs(Input::get('time')));
     Session::flash('success', 'Rank Added Successfully!');
     Redirect::to('admin.php?page=opsmanage&section=ranks');
 } elseif (Input::get('action') === 'editrank') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     try {
         Rank::update(Input::get('id'), array(
             'name' => Input::get('name'),
@@ -262,6 +341,11 @@ if (Input::get('action') === 'editprofile') {
         echo '<option>'.$name.'</option>';
     }
 } elseif (Input::get('action') === 'setcolour') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     if (!Config::replaceColour(Input::get('hexcol'))) {
         Session::flash('error', 'There was an Error Updating the Colour Theme!');
         Redirect::to('admin.php?page=site');
@@ -270,6 +354,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'Colour Theme Updated Successfully! You may need to reload the page or clear your cache in order for it to show.');
     Redirect::to('admin.php?page=site');
 } elseif (Input::get('action') === 'vasettingsupdate') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+
     if (!Config::replace('name', Input::get('vaname')) || !Config::replace('identifier', Input::get('vaident'))) {
         Session::flash('error', 'There was an error updating the Config File!');
         Redirect::to('admin.php?page=site');
@@ -278,6 +367,11 @@ if (Input::get('action') === 'editprofile') {
     Session::flash('success', 'VA Details Changed Successfully. You may need to reload the page a few times or clear your cache in order for it to show.');
     Redirect::to('admin.php?page=site');
 } elseif (Input::get('action') === 'vanetupdate') {
+    if (!$user->hasPermission('opsmanage')) {
+        Redirect::to('home.php');
+        die();
+    }
+    
     if (!Config::replace('api_key', Input::get('vanetkey'))) {
         Session::flash('error', 'There was an error updating the config file!');
         Redirect::to('admin.php?page=site');
